Renaming all gazebo_ros_control stuff to be in the same package
